+++ crates/biome_css_analyze/src/lint/nursery.rs
@@ -3,7 +3,9 @@
 use biome_analyze::declare_group;
 
 pub mod no_color_invalid_hex;
+pub mod no_css_empty_block;
 pub mod no_duplicate_font_names;
+pub mod no_duplicate_selectors_keyframe_block;
 pub mod no_unknown_unit;
 
 declare_group! {
@@ -11,7 +13,9 @@ declare_group! {
         name : "nursery" ,
         rules : [
             self :: no_color_invalid_hex :: NoColorInvalidHex ,
+            self :: no_css_empty_block :: NoCssEmptyBlock ,
             self :: no_duplicate_font_names :: NoDuplicateFontNames ,
+            self :: no_duplicate_selectors_keyframe_block :: NoDuplicateSelectorsKeyframeBlock ,
             self :: no_unknown_unit :: NoUnknownUnit ,
         ]
      }