+++ crates/biome_css_formatter/src/css/auxiliary/rule_block.rs
@@ -1,5 +1,6 @@
 use crate::prelude::*;
-use crate::utils::block_like::CssBlockLike;
+use crate::utils::block_like::FormatCssBlockLike;
+use biome_css_syntax::stmt_ext::CssBlockLike;
 use biome_css_syntax::CssRuleBlock;
 use biome_formatter::write;
 
@@ -7,7 +8,10 @@ use biome_formatter::write;
 pub(crate) struct FormatCssRuleBlock;
 impl FormatNodeRule<CssRuleBlock> for FormatCssRuleBlock {
     fn fmt_fields(&self, node: &CssRuleBlock, f: &mut CssFormatter) -> FormatResult<()> {
-        write!(f, [CssBlockLike::from(node.clone())])
+        write!(
+            f,
+            [FormatCssBlockLike::new(&CssBlockLike::from(node.clone()))]
+        )
     }
 
     fn fmt_dangling_comments(&self, _: &CssRuleBlock, _: &mut CssFormatter) -> FormatResult<()> {